+++ services/ui-src/src/utils/index.ts
@@ -23,6 +23,7 @@ export * from "./validation/validation";
 export * from "./other/email";
 export * from "./other/mask";
 export * from "./other/parsing";
+export * from "./other/print";
 export * from "./other/scrollToTop";
 export * from "./other/time";
 export * from "./other/useBreakpoint";
